+++ bcipy/display/demo/matrix/demo_matrix_layout.py
@@ -5,16 +5,18 @@ from psychopy import core
 import bcipy.display.components.layout as layout
 from bcipy.display import (InformationProperties, StimuliProperties,
                            init_display_window)
-from bcipy.display.components.task_bar import CalibrationTaskBar
+from bcipy.display.components.task_bar import CopyPhraseTaskBar
 from bcipy.display.paradigm.matrix.display import MatrixDisplay
-from bcipy.helpers.symbols import qwerty_order
 
+# from bcipy.helpers.symbols import qwerty_order
+
+font = 'Overpass Mono Medium'
 # Initialize Stimulus
 window_parameters = {
-    'full_screen': False,
+    'full_screen': True,
     'window_height': 500,
     'window_width': 700,
-    'stim_screen': 0,
+    'stim_screen': 1,
     'background_color': 'black'
 }
 
@@ -22,21 +24,24 @@ experiment_clock = core.Clock()
 win = init_display_window(window_parameters)
 win.recordFrameIntervals = False
 
-task_bar = CalibrationTaskBar(win,
-                              inquiry_count=5,
-                              current_index=0,
-                              font='Arial')
+task_bar = CopyPhraseTaskBar(win,
+                             task_text='HELLO_WORLD',
+                             spelled_text='HELLO',
+                             font=font, colors=['green'],
+                             height=0.1, padding=0.15)
 
-stim_properties = StimuliProperties(stim_font='Arial',
+matrix_height_pct = 1 - (2 * task_bar.height_pct)
+
+stim_properties = StimuliProperties(stim_font=font,
                                     stim_pos=[],
-                                    stim_height=0.17,
+                                    stim_height=0.5,
                                     is_txt_stim=True)
 
 info = InformationProperties(
     info_color=['white'],
     info_pos=[layout.at_bottom(win, height=0.21).center],
     info_height=[0.1],
-    info_font=['Arial'],
+    info_font=[font],
     info_text=['Matrix Calibration Demo'],
 )
 
@@ -45,10 +50,11 @@ matrix_display = MatrixDisplay(win,
                                stim_properties,
                                task_bar=task_bar,
                                info=info,
-                               rows=3,
-                               columns=10,
-                               width_pct=0.9,
-                               sort_order=qwerty_order(is_txt_stim=True))
+                               rows=4,
+                               columns=7,
+                               width_pct=0.7,
+                               height_pct=matrix_height_pct)
+#    sort_order=qwerty_order(is_txt_stim=True))
 
 matrix_display.draw(grid_opacity=matrix_display.full_grid_opacity,
                     grid_color=matrix_display.grid_color,