+++ bcipy/display/paradigm/vep/layout.py
@@ -2,7 +2,7 @@
 
 from itertools import cycle
 from math import sqrt
-from typing import List, NamedTuple, Optional, Tuple
+from typing import List, NamedTuple, Tuple
 
 from bcipy.display.components.layout import (Layout, above, below, left_of,
                                              right_of)
@@ -60,7 +60,7 @@ def checkerboard(squares: int, colors: Tuple[str, str], center: Tuple[float,
     center_x, center_y = center
 
     # find the center position of the left_top square
-    move_count = int(boxes_per_row / 2)
+    move_count = float(int(boxes_per_row / 2))
     if squares % 2 == 0:
         # with an even number the center is on a vertex; adjust by half a box
         move_count -= 0.5
@@ -89,11 +89,11 @@ def checkerboard(squares: int, colors: Tuple[str, str], center: Tuple[float,
 class BoxConfiguration():
     """Computes box size and positions for a VEP display.
 
-    In this configuration, there is one row on top and one on the bottom.
+    In this configuration, there are 6 areas configured with one row of 3 on top
+    and one row on the bottom.
 
     Parameters
     ----------
-        num_boxes - number of boxes; currently supports 4 or 6.
         layout - defines the boundaries within a Window in which the text areas
             will be placed
         spacing_pct - used to specify spacing between boxes within a row.
@@ -102,30 +102,25 @@ class BoxConfiguration():
 
     def __init__(self,
                  layout: Layout,
-                 num_boxes: int,
-                 spacing_pct: Optional[float] = None,
+                 spacing_pct: float = 0.05,
                  height_pct: float = 0.25):
         self.layout = layout
-        self.num_boxes = num_boxes
         self.height_pct = height_pct
         self.validate()
 
-        default_spacing = {4: 0.1, 6: 0.05}
-        if not spacing_pct:
-            spacing_pct = default_spacing[num_boxes]
         self.spacing_pct = spacing_pct
+        self._num_boxes = 6
         self._row_count = 2
 
     def validate(self):
         """Validate invariants"""
-        assert self.num_boxes == 4 or self.num_boxes == 6, 'Number of boxes must be 4 or 6'
         assert self.height_pct <= 0.5, "Rows can't take more than 50% of the height"
 
     def _box_size(self, validate: bool = True) -> Tuple[float, float]:
         """Computes the size of each box"""
         if validate:
             self.validate()
-        number_per_row = self.num_boxes / self._row_count
+        number_per_row = self._num_boxes / self._row_count
 
         # left and right boxes go to the edges, with a space between each box
         spaces_per_row = number_per_row - 1
@@ -137,6 +132,11 @@ class BoxConfiguration():
         height = self.layout.height * self.height_pct
 
         return (width, height)
+
+    @property
+    def num_boxes(self) -> int:
+        """The number of boxes"""
+        return self._num_boxes
 
     @property
     def box_size(self) -> Tuple[float, float]:
@@ -165,9 +165,30 @@ class BoxConfiguration():
         left = right_of(layout.left, width / 2)
         right = left_of(layout.right, width / 2)
 
-        positions = [(left, top), (right, top), (left, bottom),
-                     (right, bottom)]
-        if self.num_boxes == 6:
-            positions += [(layout.horizontal_middle, top),
-                          (layout.horizontal_middle, bottom)]
-        return positions
+        middle = layout.horizontal_middle
+        return [(left, top), (middle, top), (right, top), (left, bottom),
+                (middle, bottom), (right, bottom)]
+
+    @property
+    def names(self) -> List[str]:
+        """Box names for each position."""
+        return [
+            'LEFT_TOP', 'MIDDLE_TOP', 'RIGHT_TOP', 'LEFT_BOTTOM',
+            'MIDDLE_BOTTOM', 'RIGHT_BOTTOM'
+        ]
+
+
+def animation_path(start_pos: Tuple[float, float], end_pos: Tuple[float,
+                                                                  float],
+                   frames: int) -> List[Tuple[float, float]]:
+    """Computes the animation path from start_pos to end_pos"""
+    x1, y1 = start_pos
+    x2, y2 = end_pos
+    xdiff = x2 - x1
+    ydiff = y2 - y1
+    positions = []
+    for i in range(frames):
+        from_start = (i + 1) / frames
+        pos = (x1 + (from_start * xdiff), y1 + (from_start * ydiff))
+        positions.append(pos)
+    return positions