+++ bcipy/helpers/stimuli.py
@@ -8,20 +8,20 @@ from abc import ABC, abstractmethod
 from collections import Counter
 from enum import Enum
 from os import path, sep
-from typing import Any, Dict, Iterator, List, NamedTuple, Optional, Tuple, Union
+from typing import (Any, Dict, Iterator, List, NamedTuple, Optional, Tuple,
+                    Union)
 
 import mne
 import numpy as np
 import sounddevice as sd
 import soundfile as sf
-
 from mne import Annotations, Epochs
 from mne.io import RawArray
 from pandas import Series
 from PIL import Image
 from psychopy import core
 
-from bcipy.config import DEFAULT_TEXT_FIXATION, DEFAULT_FIXATION_PATH
+from bcipy.config import DEFAULT_FIXATION_PATH, DEFAULT_TEXT_FIXATION
 from bcipy.helpers.exceptions import BciPyCoreException
 from bcipy.helpers.list import grouper
 
@@ -85,7 +85,7 @@ class InquirySchedule(NamedTuple):
     - durations: `List[List[float]]`
     - colors: `List[List[str]]`
     """
-    stimuli: Union[List[List[str]], List[str]]
+    stimuli: List[Any]
     durations: Union[List[List[float]], List[float]]
     colors: Union[List[List[str]], List[str]]
 
@@ -1144,43 +1144,3 @@ def get_fixation(is_txt: bool) -> str:
         return DEFAULT_TEXT_FIXATION
     else:
         return DEFAULT_FIXATION_PATH
-
-
-def ssvep_to_code(refresh_rate: int = 60, flicker_rate: int = 10) -> List[int]:
-    """Convert SSVEP to Code.
-
-    Converts a SSVEP (steady state visual evoked potential; ex. 10 Hz) to a code (0,1)
-    given the refresh rate of the monitor (Hz) provided and a desired flicker rate (Hz).
-
-    Parameters:
-    -----------
-        refresh_rate: int, refresh rate of the monitor (Hz)
-        flicker_rate: int, desired flicker rate (Hz)
-    Returns:
-    --------
-        list of 0s and 1s that represent the code for the SSVEP on the monitor.
-    """
-    if flicker_rate > refresh_rate:
-        raise BciPyCoreException('flicker rate cannot be greater than refresh rate')
-    if flicker_rate <= 1:
-        raise BciPyCoreException('flicker rate must be greater than 1')
-
-    # get the number of frames per flicker
-    length_flicker = refresh_rate / flicker_rate
-
-    if length_flicker.is_integer():
-        length_flicker = int(length_flicker)
-    else:
-        err_message = f'flicker rate={flicker_rate} is not an integer multiple of refresh rate={refresh_rate}'
-        log.exception(err_message)
-        raise BciPyCoreException(err_message)
-
-    # start the first frames as off (0) for length of flicker;
-    # it will then toggle on (1)/ off (0) for length of flicker until all frames are filled for refresh rate.
-    t = 0
-    codes = []
-    for _ in range(flicker_rate):
-        codes += [t] * length_flicker
-        t = 1 - t
-
-    return codes