+++ bcipy/task/paradigm/matrix/copy_phrase.py
@@ -42,7 +42,8 @@ class MatrixCopyPhraseTask(RSVPCopyPhraseTask):
                             self.experiment_clock, self.spelled_text)
 
 
-def init_display(parameters, win, experiment_clock, starting_spelled_text) -> MatrixDisplay:
+def init_display(
+        parameters, win, experiment_clock, starting_spelled_text) -> MatrixDisplay:
     """Constructs a new Matrix display"""
 
     info = InformationProperties(
@@ -65,7 +66,8 @@ def init_display(parameters, win, experiment_clock, starting_spelled_text) -> Ma
                                  spelled_text=starting_spelled_text,
                                  colors=[parameters['task_color']],
                                  font=parameters['font'],
-                                 height=parameters['task_height'])
+                                 height=parameters['task_height'],
+                                 padding=parameters['task_padding'])
 
     return MatrixDisplay(
         win,
@@ -76,5 +78,6 @@ def init_display(parameters, win, experiment_clock, starting_spelled_text) -> Ma
         rows=parameters['matrix_rows'],
         columns=parameters['matrix_columns'],
         width_pct=parameters['matrix_width'],
+        height_pct=1 - (2 * task_bar.height_pct),
         trigger_type=parameters['trigger_type'],
         should_prompt_target=False)